Move tests to ReactCompositeComponent-test
